ipset/ipset.go (#4)
* ipset/ipset.go

* fixed DestroyAll function so as to not break the API

The DestroyAll() function now works with no parameters again, in case
there was another consumer of the API, we do not want to break the API
contract. So I have used a variadic trick to allow the DestroyAll()
function to accept 0 or 1 arguments (or many, but all arguments after
the first one is ignored.

I also fixed up the list() fuction in this commit since it was not
listing all names as I expected, I misunderstood how it operated before
now, so I added a new private function listAllSetNames() to get the full
list of all current ipset names and use that in the DestroyAll() loop.

I decided to do DestroyAll one by one in a loop because although ipset
can take a `ipset destroy` command that will destroy all, its not clear
what its behavior will be when it fails on one set of the many, i.e.
will it stop, and fail, or will it continue with the remaining sets?
Even if I determine one way to be true today, it maybe changed in a
future version of ipset.  I felt it was better to destroy each ipset
one-by-one in a loop and get the individual success or error status of
each one as well.

* Command.Output() to CombinedOutput()